requirements.txt
opencv-python
pyautogui
pygetwindow
pytesseract
pillow
pyyaml
